requirements.txt
https://storage.googleapis.com/tensorflow/linux/gpu/tensorflow-0.9.0-cp27-none-linux_x86_64.whl
sklearn
pandas
scipy
matplotlib
simplejson
pyyaml
